make_dataset/pyproject.toml
[tool.poetry]
name = "make_dataset"
version = "0.1.0"
description = "Transform Raw Data into Cleaned Dataset"
authors = ["Philip Chung <philip.f.chung@gmail.com>"]
readme = "README.md"
packages = [{include = "make_dataset"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pip = "^23.1"
pandas = "^1.5.3"
jupyter = "^1.0.0"
tqdm = "^4.65.0"
ipywidgets = "^8.0.6"
pyarrow = "^11.0.0"
zstd = "^1.5.5.1"
ipykernel = "^6.22.0"
azureml-core = "^1.50.0"
tiktoken = "^0.3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ruff = "^0.0.253"
mypy = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100

[tool.ruff]
ignore = []
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}
# Same as Black.
line-length = 100
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
pretty = true
color_output = true
show_column_numbers = true
show_error_context = true
show_error_codes = true
ignore_missing_imports = true
follow_imports = "skip"
# Warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
# Additional Checks
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
# Make sure we are using MyPy
disallow_untyped_calls = true
disallow_untyped_defs = true